Move clientBuilder out of workload/component.go
