+++ pkg/realizer/workload/component.go
@@ -20,12 +20,10 @@ import (
 
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	"github.com/vmware-tanzu/cartographer/pkg/apis/v1alpha1"
+	realizerclient "github.com/vmware-tanzu/cartographer/pkg/realizer/client"
 	"github.com/vmware-tanzu/cartographer/pkg/repository"
 	"github.com/vmware-tanzu/cartographer/pkg/templates"
 )
@@ -37,37 +35,6 @@ type ResourceRealizer interface {
 	Do(ctx context.Context, resource *v1alpha1.SupplyChainResource, supplyChainName string, outputs Outputs) (*unstructured.Unstructured, *templates.Output, error)
 }
 
-type ClientBuilder func(secret *corev1.Secret) (client.Client, error)
-
-func NewClientBuilder(restConfig *rest.Config) ClientBuilder {
-	return func(secret *corev1.Secret) (client.Client, error) {
-		config, err := AddBearerToken(secret, restConfig)
-		if err != nil {
-			return nil, fmt.Errorf("adding bearer token: %w", err)
-		}
-
-		cl, err := client.New(config, client.Options{})
-		if err != nil {
-			return nil, fmt.Errorf("creating client: %w", err)
-		}
-
-		return cl, nil
-	}
-}
-
-func AddBearerToken(secret *corev1.Secret, restConfig *rest.Config) (*rest.Config, error) {
-	tokenBytes, found := secret.Data[corev1.ServiceAccountTokenKey]
-	if !found {
-		return nil, fmt.Errorf("couldn't find service account token value")
-	}
-
-	newConfig := *restConfig
-	newConfig.BearerToken = string(tokenBytes)
-	newConfig.BearerTokenFile = ""
-
-	return &newConfig, nil
-}
-
 type resourceRealizer struct {
 	workload     *v1alpha1.Workload
 	systemRepo   repository.Repository
@@ -77,7 +44,7 @@ type resourceRealizer struct {
 type ResourceRealizerBuilder func(ctx context.Context, secret *corev1.Secret, workload *v1alpha1.Workload, systemRepo repository.Repository) (ResourceRealizer, error)
 
 //counterfeiter:generate sigs.k8s.io/controller-runtime/pkg/client.Client
-func NewResourceRealizerBuilder(repositoryBuilder repository.RepositoryBuilder, clientBuilder ClientBuilder, cache repository.RepoCache) ResourceRealizerBuilder {
+func NewResourceRealizerBuilder(repositoryBuilder repository.RepositoryBuilder, clientBuilder realizerclient.ClientBuilder, cache repository.RepoCache) ResourceRealizerBuilder {
 	return func(ctx context.Context, secret *corev1.Secret, workload *v1alpha1.Workload, systemRepo repository.Repository) (ResourceRealizer, error) {
 		workloadClient, err := clientBuilder(secret)
 		if err != nil {
